requirements.txt
numpy
pyrtlsdr
matplotlib
crccheck
nptyping
requests
pyserial
serial